Implement functionalities for pagination
- Created a trackable PageInfo object

- Passed the PageInfo object to the pagination UI component

- Implemented source codes to update and reflect page information objects upon the click of a pagination item

This was implemented to enhance user experience and enable them request subset of task instead of obtaining the entire task at once, and overwhelming their device with large or unlimited dataset.
Users can now click on the desired numbers below the contents to see see a related slice of tasks.